Extract common helpers to dedicated package
Move ~900 lines of shared helper code from controllers/common_helpers.go
and controllers/config_helpers.go into a new dedicated package at
cmd/thv-operator/pkg/controllerutil/.

The new package organizes utilities by domain:
- platform.go: Platform detection (Kubernetes vs OpenShift)
- rbac.go: RBAC resource management helpers
- resources.go: Resource builders and metadata utilities
- authz.go: Cedar authorization configuration
- oidc.go: OIDC authentication configuration
- tokenexchange.go: OAuth2 token exchange middleware
- config.go: Configuration tracking and hashing

This refactoring:
- Improves code organization and reusability
- Sets foundation for Phase 2 of generic reconciler pattern
- Maintains 100% functional equivalence (no behavior changes)
- Updates all imports to use ctrlutil alias
- Exports previously private constants (defaultAuthzKey -> DefaultAuthzKey)
- Moves tests to new package location

All existing tests pass and linter reports 0 issues.
