+++ cmd/thv-operator/controllers/mcpserver_resource_overrides_test.go
@@ -26,6 +26,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client/fake"
 
 	mcpv1alpha1 "github.com/stacklok/toolhive/cmd/thv-operator/api/v1alpha1"
+	ctrlutil "github.com/stacklok/toolhive/cmd/thv-operator/pkg/controllerutil"
 	"github.com/stacklok/toolhive/pkg/container/kubernetes"
 )
 
@@ -397,7 +398,7 @@ func TestMergeStringMaps(t *testing.T) {
 		t.Run(tt.name, func(t *testing.T) {
 			t.Parallel()
 
-			result := mergeStringMaps(tt.defaultMap, tt.overrideMap)
+			result := ctrlutil.MergeStringMaps(tt.defaultMap, tt.overrideMap)
 			assert.Equal(t, tt.expected, result)
 		})
 	}
