+++ cmd/thv-operator/pkg/controllerutil/config_test.go
@@ -1,4 +1,4 @@
-package controllers
+package controllerutil
 
 import (
 	"testing"
