+++ cmd/thv-operator/test-integration/mcp-server/suite_test.go
@@ -24,6 +24,7 @@ import (
 
 	mcpv1alpha1 "github.com/stacklok/toolhive/cmd/thv-operator/api/v1alpha1"
 	"github.com/stacklok/toolhive/cmd/thv-operator/controllers"
+	ctrlutil "github.com/stacklok/toolhive/cmd/thv-operator/pkg/controllerutil"
 )
 
 // These tests use Ginkgo (BDD-style Go testing framework). Refer to
@@ -93,7 +94,7 @@ var _ = BeforeSuite(func() {
 	err = (&controllers.MCPServerReconciler{
 		Client:           k8sManager.GetClient(),
 		Scheme:           k8sManager.GetScheme(),
-		PlatformDetector: controllers.NewSharedPlatformDetector(),
+		PlatformDetector: ctrlutil.NewSharedPlatformDetector(),
 	}).SetupWithManager(k8sManager)
 	Expect(err).ToNot(HaveOccurred())
